refactoring the footer component
