+++ components/AppRouterMigrationComponents/footer/constants.ts
@@ -1,0 +1,168 @@
+export const footerNavZh = [
+  {
+    header: '产品',
+    footerItem: [
+      {
+        href: '/zh/showcase',
+        label: '案例展示',
+      },
+      {
+        href: 'https://app.tina.io',
+        label: 'TinaCloud',
+      },
+      {
+        href: '/docs',
+        label: '产品介绍',
+      },
+      {
+        href: '/docs/product-tour',
+        label: '工作原理',
+      },
+      {
+        label: '产品路线',
+        href: '/zh/roadmap',
+      },
+    ],
+  },
+  {
+    header: '资源',
+    footerItem: [
+      {
+        label: '博客',
+        href: '/blog',
+      },
+      {
+        label: '示例',
+        href: '/examples',
+      },
+      {
+        label: '产品对比',
+        href: '/zh/compare-tina',
+      },
+      {
+        label: '技术支持',
+        href: '/docs/support',
+      },
+      {
+        href: '/media',
+        label: '媒体资源',
+      },
+    ],
+  },
+  {
+    label: '',
+    footerItem: [
+      {
+        label: '最新动态',
+        items: [
+          {
+            href: '/whats-new/tinacms',
+            label: 'TinaCMS',
+          },
+          {
+            href: '/whats-new/tinacloud',
+            label: 'TinaCloud',
+          },
+        ],
+      },
+      {
+        label: '应用场景',
+        items: [
+          {
+            href: '/agencies',
+            label: '开发机构',
+          },
+          {
+            href: '/documentation',
+            label: '文档管理',
+          },
+          {
+            href: '/cms-for-teams',
+            label: '团队协作',
+          },
+          {
+            href: '/jamstack-cms',
+            label: 'Jamstack CMS',
+          },
+        ],
+      },
+      {
+        label: '核心优势',
+        items: [
+          {
+            href: '/mdx-cms',
+            label: 'MDX支持',
+          },
+          {
+            href: '/markdown-cms',
+            label: 'Markdown支持',
+          },
+          {
+            href: '/git-cms',
+            label: 'Git集成',
+          },
+          {
+            href: '/editorial-workflow',
+            label: '编辑工作流',
+          },
+          {
+            href: '/flexible-cms',
+            label: '高度定制',
+          },
+          {
+            href: '/seo',
+            label: 'SEO优化',
+          },
+        ],
+      },
+      {
+        label: '集成方案',
+        items: [
+          {
+            href: '/astro',
+            label: 'Astro',
+          },
+          {
+            href: '/hugo-cms',
+            label: 'Hugo',
+          },
+          {
+            href: '/nextjs-cms',
+            label: 'NextJS',
+          },
+          {
+            href: '/jekyll-cms',
+            label: 'Jekyll',
+          },
+        ],
+      },
+    ],
+  },
+] as const;
+
+export const footerLinksZh = [
+  {
+    href: '/security',
+    label: '安全政策',
+  },
+  {
+    href: '/telemetry',
+    label: '开源遥测',
+  },
+  {
+    href: '/terms-of-service',
+    label: '服务条款',
+  },
+  {
+    href: '/privacy-notice',
+    label: '隐私声明',
+  },
+  {
+    href: 'https://github.com/tinacms/tinacms/blob/master/LICENSE',
+    label: '开源许可',
+  },
+  {
+    href: '/docs/support',
+    label: '技术支持',
+  },
+];
